Move v8 Rating changes to from react-next to react
