--- apps/perf-test/src/scenarios/RatingNext.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react';
-import { Rating } from '@fluentui/react-next';
-
-const Scenario = () => <Rating />;
-
-export default Scenario;
