--- packages/react-examples/src/react-next/Rating/docs/RatingBestPractices.md
@@ -1,9 +1,0 @@
-### Layout
-
-- Make it clear which item the rating pertains to by making sure the layout and grouping are clear when several items are on the page.
-- Don't use the rating component for data that has a continuous range, such as the brightness of a photo. Instead, use a slider.
-
-### Content
-
-- Use a five-star rating system.
-- Use sentence-style capitalization—only capitalize the first word. For more info, see [Capitalization](https://docs.microsoft.com/style-guide/capitalization) in the Microsoft Writing Style Guide.
