--- packages/react-examples/src/react-next/Rating/docs/RatingOverview.md
@@ -1,1 +1,0 @@
-Ratings show people’s opinions of a product, helping others make more informed purchasing decisions. People can also rate products they’ve purchased.
