+++ packages/react-examples/src/react/Rating/Rating.ButtonControlled.Example.tsx
@@ -1,18 +1,14 @@
 import * as React from 'react';
-import { Rating } from '@fluentui/react/lib/Rating';
-import { PrimaryButton } from '@fluentui/react/lib/Button';
+import { Rating } from '@fluentui/react';
+import { PrimaryButton } from '@fluentui/react/lib/compat/Button';
 
 export const RatingButtonControlledExample: React.FunctionComponent = () => {
   const [currentRating, setCurrentRating] = React.useState(5);
-  const currentRatingToggle = () => {
-    if (currentRating === 0) {
-      setCurrentRating(5);
-    } else {
-      setCurrentRating(0);
-    }
-  };
+  const currentRatingToggle = React.useCallback(() => {
+    setCurrentRating(oldRating => (oldRating === 0 ? 5 : 0));
+  }, []);
 
-  const getRatingAriaLabel = () => `Rating value is ${currentRating} of 5`;
+  const getRatingAriaLabel = React.useCallback(() => `Rating value is ${currentRating} of 5`, [currentRating]);
 
   return (
     <div>
@@ -21,15 +17,10 @@ export const RatingButtonControlledExample: React.FunctionComponent = () => {
         max={5}
         readOnly
         allowZeroStars
-        // eslint-disable-next-line react/jsx-no-bind
         getAriaLabel={getRatingAriaLabel}
-        ariaLabelFormat={'Select {0} of {1} stars'}
+        ariaLabelFormat="Select {0} of {1} stars"
       />
-      <PrimaryButton
-        text={'Click to change rating to ' + (5 - currentRating)}
-        // eslint-disable-next-line react/jsx-no-bind
-        onClick={currentRatingToggle}
-      />
+      <PrimaryButton onClick={currentRatingToggle}>Click to change rating to {5 - currentRating}</PrimaryButton>
     </div>
   );
 };
