--- packages/react-next/src/components/Rating/index.ts
@@ -1,3 +1,0 @@
-export * from './Rating';
-export * from './Rating.base';
-export * from './Rating.types';
